requirements.txt
flask
dotenv
bcrypt
mysql.connector